add: chainid to feed
